move navbar to Navbar component, move tabs to Tabs component
