+++ src/components/Tabs/Tabs.tsx
@@ -1,20 +1,18 @@
+import React from 'react';
+import { Tab } from '../../types/Tab';
 import classNames from 'classnames';
 import { Link, useParams } from 'react-router-dom';
 
-const tabs = [
-  { id: 'tab-1', title: 'Tab 1', content: 'Some text 1' },
-  { id: 'tab-2', title: 'Tab 2', content: 'Some text 2' },
-  { id: 'tab-3', title: 'Tab 3', content: 'Some text 3' },
-];
+type Props = {
+  tabs: Tab[];
+};
 
-export const TabsPage: React.FC = () => {
+export const Tabs: React.FC<Props> = ({ tabs }) => {
   const { tabId } = useParams();
   const activeTab = tabs.find(tab => tab.id === tabId);
 
   return (
     <>
-      <h1 className="title">Tabs page</h1>
-
       <div className="tabs is-boxed">
         <ul>
           {tabs.map(tab => (
