+++ src/components/Tabs/index.ts
@@ -1,0 +1,1 @@
+export * from './Tabs';
